Android calculator update 3
Completely changed the way numbers are calculated, expressions are now
converted to postfix notation and put onto a binary tree that can
efficiently operate on all of the terms. This is a much more elegant way
to calculate the numbers and makes it a lot easier to add new features
in later (exponents, parentheses, basically any other type of ordered
operation). Also updated error handling again to only target specific
runtime errorrs that can be caused by bad user input.